mysite/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = "Django shop project"
authors = ["Your Name <youremail@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.2.6"
djangorestframework = "3.16.1"
django-filter = "25.2"
django-redis = "6.0.0"
drf-spectacular = "0.28.0"
pillow = "11.3.0"
PyYAML = "6.0.3"
redis = "6.4.0"
gunicorn = ">=20.0.4"
django-debug-toolbar = "6.0.0"
inflection = "0.5.1"
jsonschema = "4.25.1"
jsonschema-specifications = "2025.9.1"
referencing = "0.37.0"
rpds-py = "0.28.0"
sqlparse = "0.5.3"
uritemplate = "4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
